// frontend/public/src/data-utils/models/Enums.ts
export enum ItemSubTypeEnum {
  None = 0,
  Crucible = 1,
  Vanguard = 2,
  Exotic = 5,
  AutoRifle = 6,
  Shotgun = 7,
  Machinegun = 8,
  HandCannon = 9,
  RocketLauncher = 10,
  FusionRifle = 11,
  SniperRifle = 12,
  PulseRifle = 13,
  ScoutRifle = 14,
  Crm = 16,
  Sidearm = 17,
  Sword = 18,
  Mask = 19,
  Shader = 20,
  Ornament = 21,
  FusionRifleLine = 22,
  GrenadeLauncher = 23,
  SubmachineGun = 24,
  TraceRifle = 25,
  HelmetArmor = 26,
  GauntletsArmor = 27,
  ChestArmor = 28,
  LegArmor = 29,
  ClassArmor = 30,
  Bow = 31,
  DummyRepeatableBounty = 32
}

export enum ClassEnum {
  Titan = 0,
  Hunter = 1,
  Warlock = 2,
  Unknown = 3
}

export enum EnergyTypeEnum {
  Any = 0,
  Solar = 2,
  Arc = 1,
  Void = 3,
  Ghost = 4,
  Subclass = 5,
  Stasis = 6,
}

export enum DamageTypeEnum {
  None = 0,
  Kinetic = 1,
  Arc = 2,
  Thermal = 3,
  Void = 4,
  Raid = 5,
  Stasis = 6
}

export enum BucketTypeEnum {
  Kinetic = 1498876634,
  Energy = 2465295065,
  Power = 953998645,
  Helmet = 3448274439,
  Arms = 3551918588,
  Chest = 14239492,
  Legs = 20886954,
  ClassItem = 1585787867,
  Subclass = 3284755031
}

export enum ItemTypeEnum {
  None = 0,
  Currency = 1,
  Armor = 2,
  Weapon = 3,
  Message = 7,
  Engram = 8,
  Consumable = 9,
  ExchangeMaterial = 10,
  MissionReward = 11,
  QuestStep = 12,
  QuestStepComplete = 13,
  Emblem = 14,
  Quest = 15,
  Subclass = 16,
  ClanBanner = 17,
  Aura = 18,
  Mod = 19,
  Dummy = 20,
  Ship = 21,
  Vehicle = 22,
  Emote = 23,
  Ghost = 24,
  Package = 25,
  Bounty = 26,
  Wrapper = 27,
  SeasonalArtifact = 28,
  Finisher = 29
}

export enum SocketTypeEnum {
  IntrinsicTraits = 3956125808,
  WeaponPerks = 4241085061,
  WeaponMods = 2685412949,
  WeaponCosmetics = 2048875504,
  ArmorMods = 590099826, // TODO: Find the right hash for this
  ArmorPerks = 3154740035
}

export enum WeaponMasterworkType {
  None = 0,
  Range = 1
}